+++ ToDoList/events.js
@@ -1,0 +1,7 @@
+/* export function deleteAllOnClick(toDos) {
+    const itemContainer = document.getElementById('item_container')
+    itemContainer.innerHTML = ''
+
+    toDos.length = 0
+    localStorage.setItem('todos', JSON.stringify(toDos))
+} */
