+++ ToDoList/index.html
@@ -14,6 +14,6 @@
     <div class="root" id="root">
        
     </div>
-    <script src='script.js'></script>
+    <script type='module' src='./index.js'></script>
 </body>
 </html>
