+++ ToDoList/index.js
@@ -1,0 +1,6 @@
+import { getItemsFromLocalStorage } from './localStorage.js'
+import { createPage } from './createLayout.js'
+
+const toDos = getItemsFromLocalStorage()
+
+createPage(toDos)
